# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DEMS API",
    "description": "Plataforma de evaluación de prototipos IPN-DEMS.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/admin/editions/{id}": {
      "get": {
        "tags": [
          "admin/editions"
        ],
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditionView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin/editions"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "Tiene rúbricas o prototipos"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "admin/editions"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEditionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditionView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/admin/prototipos/{id}": {
      "get": {
        "tags": [
          "admin/prototipos"
        ],
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototipoView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin/prototipos"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "Tiene evaluaciones"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "admin/prototipos"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchPrototipoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototipoView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/admin/rubric-templates/{id}": {
      "get": {
        "tags": [
          "admin/rubrics"
        ],
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricTemplateView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin/rubrics"
        ],
        "operationId": "delete_rubric",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "Hay evaluaciones referenciando — desactivar en su lugar"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "admin/rubrics"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRubricRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricTemplateView"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": [
          "admin/users"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "description": "Filtra por rol",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserRole"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "description": "Filtra por activo",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin/users"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Usuario creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "409": {
            "description": "Email ya registrado"
          },
          "422": {
            "description": "Validación falló"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "tags": [
          "admin/users"
        ],
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin/users"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Borrado"
          },
          "404": {
            "description": "No encontrado"
          },
          "409": {
            "description": "Tiene evaluaciones — desactivar en su lugar"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "admin/users"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/admin/users/{id}/password": {
      "put": {
        "tags": [
          "admin/users"
        ],
        "operationId": "reset_password",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del usuario",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password reseteada"
          },
          "404": {
            "description": "No encontrado"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Nuevos tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido o usuario inactivo"
          }
        }
      }
    },
...
    "/evaluaciones/{id}": {
      "get": {
        "tags": [
          "evaluaciones"
        ],
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la evaluación",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evaluación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluacionView"
                }
              }
            }
          },
          "403": {
            "description": "No es el dueño ni admin"
          },
          "404": {
            "description": "No encontrada"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "evaluaciones"
        ],
        "operationId": "patch_evaluacion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la evaluación",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEvaluacionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Evaluación actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluacionView"
                }
              }
            }
          },
          "403": {
            "description": "No es el dueño"
          },
          "404": {
            "description": "No encontrada"
          },
          "409": {
            "description": "Conflicto. Body lleva `code`: `already_submitted` (ya enviada — chequeado antes que la fase) o `edition_closed` (edición fuera de fase 'evaluacion')."
          },
          "422": {
            "description": "Body inválido"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "email",
          "full_name",
          "role",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        }
      },
...
      "EditionView": {
        "type": "object",
        "required": [
          "id",
          "year",
          "name",
          "active",
          "phase",
          "created_at"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "phase": {
            "$ref": "#/components/schemas/EditionPhase"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PatchEditionRequest": {
        "type": "object",
        "properties": {
          "active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PatchPrototipoRequest": {
        "type": "object",
        "properties": {
          "descripcion": {
            "type": [
              "string",
              "null"
            ]
          },
          "eje_transversal": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "nombre": {
            "type": [
              "string",
              "null"
            ]
          },
          "plantel": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "PatchRubricRequest": {
        "type": "object",
        "properties": {
          "activo": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "descripcion": {
            "type": [
              "string",
              "null"
            ]
          },
          "nombre": {
            "type": [
              "string",
              "null"
            ]
          },
          "peso": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Peso porcentual (0..100). Si se omite, no se toca."
          }
        }
      },
      "PatchUserRequest": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "role": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          }
        }
      },
...
      "PrototipoView": {
        "type": "object",
        "required": [
          "id",
          "edition_id",
          "folio",
          "nombre",
          "eje_transversal",
          "categorias",
          "integrantes",
          "created_at"
        ],
        "properties": {
          "categorias": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "descripcion": {
            "type": [
              "string",
              "null"
            ]
          },
          "edition_id": {
            "type": "string",
            "format": "uuid"
          },
          "eje_transversal": {
            "type": "boolean"
          },
          "folio": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "integrantes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntegranteView"
            }
          },
          "nombre": {
            "type": "string"
          },
          "plantel": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "ResetPasswordRequest": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
      "RubricTemplateView": {
        "type": "object",
        "required": [
          "id",
          "edition_id",
          "nombre",
          "tipo",
          "activo",
          "peso",
          "editable",
          "categorias",
          "sections"
        ],
        "properties": {
          "activo": {
            "type": "boolean"
          },
          "categorias": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "descripcion": {
            "type": [
              "string",
              "null"
            ]
          },
          "editable": {
            "type": "boolean",
            "description": "`true` si la edición está en fase `preparacion`: la estructura puede\neditarse. `false` cuando la edición ya entró en evaluación/cerrada."
          },
          "edition_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "peso": {
            "type": "integer",
            "format": "int32",
            "description": "Peso porcentual (0..100) de esta rúbrica en el puntaje final combinado\ndel prototipo."
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionView"
            }
          },
          "tipo": {
            "$ref": "#/components/schemas/RubricType"
          }
        }
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "jurado"
        ]
      },
      "UserView": {
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_active": {
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}